+++ packages/stream_chat/lib/src/core/api/responses.g.dart
@@ -476,8 +476,31 @@ GetUnreadCountResponse _$GetUnreadCountResponseFromJson(
           .map((e) => UnreadCountsThread.fromJson(e as Map<String, dynamic>))
           .toList();
 
-GetActiveLiveLocationsResponse _$GetActiveLiveLocationsResponseFromJson(
+UpsertPushPreferencesResponse _$UpsertPushPreferencesResponseFromJson(
         Map<String, dynamic> json) =>
+    UpsertPushPreferencesResponse()
+      ..duration = json['duration'] as String?
+      ..userPreferences = (json['user_preferences'] as Map<String, dynamic>?)
+              ?.map(
+            (k, e) =>
+                MapEntry(k, PushPreference.fromJson(e as Map<String, dynamic>)),
+          ) ??
+          {}
+      ..userChannelPreferences =
+          (json['user_channel_preferences'] as Map<String, dynamic>?)?.map(
+                (k, e) => MapEntry(
+                    k,
+                    (e as Map<String, dynamic>).map(
+                      (k, e) => MapEntry(
+                          k,
+                          ChannelPushPreference.fromJson(
+                              e as Map<String, dynamic>)),
+                    )),
+              ) ??
+              {};
+
+GetActiveLiveLocationsResponse _$GetActiveLiveLocationsResponseFromJson(
+    Map<String, dynamic> json) =>
     GetActiveLiveLocationsResponse()
       ..duration = json['duration'] as String?
       ..activeLiveLocations = (json['active_live_locations'] as List<dynamic>)