+++ packages/stream_chat/lib/src/core/models/channel_state.dart
@@ -5,6 +5,7 @@ import 'package:stream_chat/src/core/models/draft.dart';
 import 'package:stream_chat/src/core/models/location.dart';
 import 'package:stream_chat/src/core/models/member.dart';
 import 'package:stream_chat/src/core/models/message.dart';
+import 'package:stream_chat/src/core/models/push_preference.dart';
 import 'package:stream_chat/src/core/models/read.dart';
 import 'package:stream_chat/src/core/models/user.dart';
 
@@ -20,7 +21,7 @@ const _nullConst = _NullConst();
 @JsonSerializable()
 class ChannelState implements ComparableFieldProvider {
   /// Constructor used for json serialization
-  ChannelState({
+  const ChannelState({
     this.channel,
     this.messages,
     this.members,
@@ -30,6 +31,7 @@ class ChannelState implements ComparableFieldProvider {
     this.read,
     this.membership,
     this.draft,
+    this.pushPreferences,
     this.activeLiveLocations,
   });
 
@@ -60,6 +62,9 @@ class ChannelState implements ComparableFieldProvider {
   /// The draft message for this channel if it exists.
   final Draft? draft;
 
+  /// The push preferences for this channel if it exists.
+  final ChannelPushPreference? pushPreferences;
+
   /// The list of active live locations in the channel.
   final List<Location>? activeLiveLocations;
 
@@ -81,6 +86,7 @@ class ChannelState implements ComparableFieldProvider {
     List<Read>? read,
     Member? membership,
     Object? draft = _nullConst,
+    ChannelPushPreference? pushPreferences,
     List<Location>? activeLiveLocations,
   }) =>
       ChannelState(
@@ -93,6 +99,7 @@ class ChannelState implements ComparableFieldProvider {
         read: read ?? this.read,
         membership: membership ?? this.membership,
         draft: draft == _nullConst ? this.draft : draft as Draft?,
+        pushPreferences: pushPreferences ?? this.pushPreferences,
         activeLiveLocations: activeLiveLocations ?? this.activeLiveLocations,
       );