+++ packages/stream_chat/lib/src/core/models/channel_state.g.dart
@@ -32,9 +32,13 @@ ChannelState _$ChannelStateFromJson(Map<String, dynamic> json) => ChannelState(
       draft: json['draft'] == null
           ? null
           : Draft.fromJson(json['draft'] as Map<String, dynamic>),
-      activeLiveLocations: (json['active_live_locations'] as List<dynamic>?)
-          ?.map((e) => Location.fromJson(e as Map<String, dynamic>))
-          .toList(),
+      pushPreferences: json['push_preferences'] == null
+          ? null
+          : ChannelPushPreference.fromJson(
+              json['push_preferences'] as Map<String, dynamic>),
+  activeLiveLocations: (json['active_live_locations'] as List<dynamic>?)
+      ?.map((e) => Location.fromJson(e as Map<String, dynamic>))
+      .toList(),
     );
 
 Map<String, dynamic> _$ChannelStateToJson(ChannelState instance) =>
@@ -49,6 +53,7 @@ Map<String, dynamic> _$ChannelStateToJson(ChannelState instance) =>
       'read': instance.read?.map((e) => e.toJson()).toList(),
       'membership': instance.membership?.toJson(),
       'draft': instance.draft?.toJson(),
+      'push_preferences': instance.pushPreferences?.toJson(),
       'active_live_locations':
-          instance.activeLiveLocations?.map((e) => e.toJson()).toList(),
+      instance.activeLiveLocations?.map((e) => e.toJson()).toList(),
     };