+++ packages/stream_chat/test/src/db/chat_persistence_client_test.dart
@@ -268,7 +268,7 @@ void main() {
     });
 
     test('updateChannelState', () async {
-      final channelState = ChannelState();
+      const channelState = ChannelState();
       persistenceClient.updateChannelState(channelState);
     });