+++ packages/stream_chat_flutter/lib/src/message_list_view/message_list_view.dart
@@ -462,10 +462,6 @@ class _StreamMessageListViewState extends State<StreamMessageListView> {
           _firstUnreadMessage = streamChannel?.getFirstUnreadMessage();
         }),
       );
-
-      if (_isThreadConversation) {
-        streamChannel!.getReplies(widget.parentMessage!.id);
-      }
     }
   }