+++ packages/stream_chat_persistence/lib/src/dao/message_dao.dart
@@ -157,9 +157,9 @@ class MessageDao extends DatabaseAccessor<DriftChatDatabase>
           msgList.removeRange(lessThanIndex, msgList.length);
         }
       }
-      if (options?.greaterThanOrEqual != null) {
+      if (options?.greaterThan != null) {
         final greaterThanIndex = msgList.indexWhere(
-          (m) => m.id == options!.greaterThanOrEqual,
+          (m) => m.id == options!.greaterThan,
         );
         if (greaterThanIndex != -1) {
           msgList.removeRange(0, greaterThanIndex);
@@ -214,9 +214,9 @@ class MessageDao extends DatabaseAccessor<DriftChatDatabase>
           msgList.removeRange(lessThanIndex, msgList.length);
         }
       }
-      if (messagePagination?.greaterThanOrEqual != null) {
+      if (messagePagination?.greaterThan != null) {
         final greaterThanIndex = msgList.indexWhere(
-          (m) => m.id == messagePagination!.greaterThanOrEqual,
+          (m) => m.id == messagePagination!.greaterThan,
         );
         if (greaterThanIndex != -1) {
           msgList.removeRange(0, greaterThanIndex);