+++ packages/stream_chat/CHANGELOG.md
@@ -1,10 +1,18 @@
-## Upcoming
+## 9.16.0
 
 🐞 Fixed
 
 - Fixed `skipPush` and `skipEnrichUrl` not preserving during message send or update retry
 - Fixed `Channel` methods to throw proper `StateError` exceptions instead of relying on assertions
   for state validation.
+- Fixed `OwnUser` specific fields getting lost when creating a new `OwnUser` instance from
+  an `User` instance.
+- Fixed `Client.currentUser` specific fields getting reset on `user.updated` events.
+
+✅ Added
+
+- Added support for `Client.setPushPreferences` which allows setting PushPreferences for the
+  current user or for a specific channel.
 
 ## 10.0.0-beta.4