+++ packages/stream_chat_localizations/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 9.16.0
+
+- Updated `stream_chat_flutter` dependency to [`9.16.0`](https://pub.dev/packages/stream_chat_flutter/changelog).
+
 ## 10.0.0-beta.4
 
 - Added translations for new `locationLabel` label.