+++ packages/stream_chat_persistence/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 9.16.0
+
+- Updated `stream_chat` dependency to [`9.16.0`](https://pub.dev/packages/stream_chat/changelog).
+
 ## 10.0.0-beta.4
 
 - Added support for `Location` entity in the database.